refactor(auth): Move from component/pages to pages root directory
